change formatting of license file
This change brings the formatting of the file as close to the
canonical text of 2-clause BSD without changing any of the actual
words. Just whitespace, newline positions and bullet point style.

It is still not 100% match because of the "CC-XJC PROJECT" change
to the original text, but it gets licensee [1] to detect it as
2-clause BSD style with 98% accuracy instead of "unknown".

This should convince GitHub to detect it correctly.

[1] https://github.com/licensee/licensee